+++ test/index.js
@@ -1,10 +1,15 @@
 const assert = require( 'assert' );
 const sinon = require( 'sinon' );
-const createSelector = require( '../' );
 
-describe( 'createSelector', () => {
+function test( createSelector, WeakMapImpl ) {
 	let getTasksByCompletion;
 	const sandbox = sinon.sandbox.create();
+
+	function ifWeakMapIt( ...args ) {
+		if ( typeof WeakMapImpl !== 'undefined' ) {
+			return it( ...args );
+		}
+	}
 
 	const selector = sandbox.spy(
 		( state, isComplete = false, extra ) => (
@@ -63,46 +68,6 @@ describe( 'createSelector', () => {
 		assert.deepEqual( completed, selector( state, true, obj ) );
 	} );
 
-	it( 'caches with maxSize', () => {
-		getTasksByCompletion = createSelector(
-			selector,
-			( state ) => state.todo,
-			{ maxSize: 2 }
-		);
-
-		const state = getState();
-		const result = getTasksByCompletion( state, true );
-		selector.reset();
-
-		// cache MISS [ [ true, 1 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 1 ), [ ...result, 1 ] );
-		sinon.assert.callCount( selector, 1 );
-
-		// cache MISS [ [ true, 2 ], [ true, 1 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 2 ), [ ...result, 2 ] );
-		sinon.assert.callCount( selector, 2 );
-
-		// cache MISS [ [ true, 3 ], [ true, 2 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 3 ), [ ...result, 3 ] );
-		sinon.assert.callCount( selector, 3 );
-
-		// cache MISS [ [ true, 1 ], [ true, 3 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 1 ), [ ...result, 1 ] );
-		sinon.assert.callCount( selector, 4 );
-
-		// cache HIT [ [ true, 3 ], [ true, 1 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 3 ), [ ...result, 3 ] );
-		sinon.assert.callCount( selector, 4 );
-
-		// cache HIT [ [ true, 1 ], [ true, 3 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 1 ), [ ...result, 1 ] );
-		sinon.assert.callCount( selector, 4 );
-
-		// cache MISS [ [ true, 2 ], [ true, 1 ] ]
-		assert.deepEqual( getTasksByCompletion( state, true, 2 ), [ ...result, 2 ] );
-		sinon.assert.callCount( selector, 5 );
-	} );
-
 	it( 'defaults to caching on entire state object', () => {
 		const getTasksByCompletionOnState = createSelector( selector );
 		let state = getState();
@@ -146,6 +111,18 @@ describe( 'createSelector', () => {
 		assert.deepEqual( completed, selector( state, true ) );
 	} );
 
+	it( 'clears cache when primitive dependant changes', () => {
+		const getTasksByCompletionOnCount = createSelector( selector, ( state ) => state.ok );
+		const state = getState();
+		state.ok = false;
+		let completed = getTasksByCompletionOnCount( state, true );
+		state.ok = true;
+		completed = getTasksByCompletionOnCount( state, true );
+
+		sinon.assert.calledTwice( selector );
+		assert.deepEqual( completed, selector( state, true ) );
+	} );
+
 	it( 'clears cache when non-primitive argument reference changes', () => {
 		const state = getState();
 		let completed;
@@ -170,6 +147,41 @@ describe( 'createSelector', () => {
 		assert.deepEqual( completed, selector( state, true ) );
 	} );
 
+	ifWeakMapIt( 'deep caches on object dependants', () => {
+		const stateByDate = {
+			todos: {
+				'2018-01-01': [
+					{ text: 'Go to the gym', complete: true },
+					{ text: 'Try to spend time in the sunlight', complete: true }
+				],
+				'2018-01-02': [
+					{ text: 'Laundry must be done', complete: false }
+				]
+			}
+		};
+
+		const selectorByDate = sandbox.spy( ( state, date, isComplete = false ) => (
+			state.todos[ date ].filter( ( task ) => task.complete === isComplete )
+		) );
+
+		const getTasksByCompletionByDate = createSelector(
+			selectorByDate,
+			( state, date ) => state.todos[ date ]
+		);
+
+		getTasksByCompletionByDate( stateByDate, '2018-01-01' );
+		sinon.assert.calledOnce( selectorByDate );
+
+		getTasksByCompletionByDate( stateByDate, '2018-01-02' );
+		sinon.assert.calledTwice( selectorByDate );
+
+		getTasksByCompletionByDate( stateByDate, '2018-01-01' );
+		sinon.assert.calledTwice( selectorByDate );
+
+		getTasksByCompletionByDate( stateByDate, '2018-01-02' );
+		sinon.assert.calledTwice( selectorByDate );
+	} );
+
 	it( 'ensures equal argument length before returning cache', () => {
 		const state = getState();
 		let completed;
@@ -179,4 +191,35 @@ describe( 'createSelector', () => {
 		sinon.assert.calledTwice( selector );
 		assert.deepEqual( completed, selector( state ) );
 	} );
+}
+
+describe( 'createSelector', () => {
+	const _WeakMap = WeakMap;
+
+	const WEAKMAP_SUPPORT = {
+		'with': _WeakMap,
+		without: undefined
+	};
+
+	for ( const name in WEAKMAP_SUPPORT ) {
+		const WeakMapImpl = WEAKMAP_SUPPORT[ name ];
+
+		context( name + ' WeakMap', () => {
+			let implementation;
+			const createSelector = ( ...args ) => implementation( ...args );
+
+			before( () => {
+				global.WeakMap = WeakMapImpl;
+
+				delete require.cache[ require.resolve( '../' ) ];
+				implementation = require( '../' );
+			} );
+
+			after( () => {
+				global.WeakMap = _WeakMap;
+			} );
+
+			test( createSelector, WeakMapImpl );
+		} );
+	}
 } );